fix (init Data): init data synchronously for each panel in config: firstly store the data in store and then show the panels
